# Explicit Enzyme rules