--- a/.github/workflows/AD.yml
+++ b/.github/workflows/AD.yml
@@ -14,6 +14,7 @@ jobs:
       matrix:
         version:
           - '1.6'
+          - '1.10'
           - '1'
         os:
           - ubuntu-latest