--- a/test/ad/enzyme.jl
+++ b/test/ad/enzyme.jl
@@ -1,34 +1,38 @@
-@testset "Enzyme: Bijectors.find_alpha" begin
-    x = randn()
-    y = expm1(randn())
-    z = randn()
+# Segfaults on older Julia versions, probably never supported
+# TODO: Enable tests on Julia >= 1.11 when updating to Enzyme 0.13
+if v"1.10" <= VERSION < v"1.11"
+    @testset "Enzyme: Bijectors.find_alpha" begin
+        x = randn()
+        y = expm1(randn())
+        z = randn()
 
-    @testset "forward" begin
-        # No batches
-        @testset for RT in (Const, Duplicated, DuplicatedNoNeed),
-            Tx in (Const, Duplicated),
-            Ty in (Const, Duplicated),
-            Tz in (Const, Duplicated)
+        @testset "forward" begin
+            # No batches
+            @testset for RT in (Const, Duplicated, DuplicatedNoNeed),
+                Tx in (Const, Duplicated),
+                Ty in (Const, Duplicated),
+                Tz in (Const, Duplicated)
 
-            test_forward(Bijectors.find_alpha, RT, (x, Tx), (y, Ty), (z, Tz))
+                test_forward(Bijectors.find_alpha, RT, (x, Tx), (y, Ty), (z, Tz))
+            end
+
+            # Batches
+            @testset for RT in (Const, BatchDuplicated, BatchDuplicatedNoNeed),
+                Tx in (Const, BatchDuplicated),
+                Ty in (Const, BatchDuplicated),
+                Tz in (Const, BatchDuplicated)
+
+                test_forward(Bijectors.find_alpha, RT, (x, Tx), (y, Ty), (z, Tz))
+            end
         end
+        @testset "reverse" begin
+            @testset for RT in (Const, Active),
+                Tx in (Const, Active),
+                Ty in (Const, Active),
+                Tz in (Const, Active)
 
-        # Batches
-        @testset for RT in (Const, BatchDuplicated, BatchDuplicatedNoNeed),
-            Tx in (Const, BatchDuplicated),
-            Ty in (Const, BatchDuplicated),
-            Tz in (Const, BatchDuplicated)
-
-            test_forward(Bijectors.find_alpha, RT, (x, Tx), (y, Ty), (z, Tz))
-        end
-    end
-    @testset "reverse" begin
-        @testset for RT in (Const, Active),
-            Tx in (Const, Active),
-            Ty in (Const, Active),
-            Tz in (Const, Active)
-
-            test_reverse(Bijectors.find_alpha, RT, (x, Tx), (y, Ty), (z, Tz))
+                test_reverse(Bijectors.find_alpha, RT, (x, Tx), (y, Ty), (z, Tz))
+            end
         end
     end
 end