@torfjelde 2024-11-11
> can you help review this PR since you worked on the rules for ReverseDiff?

I have no clue about Enzyme.jl, unfortunately 😕  Seems @wsmoses's feedback is much more useful 👍