@devmotion 2024-11-28
Superseded and fixed by #350.